requirements.txt
numpy
opencv-python
pillow
torch
torchvision
tensorboard
matplotlib
scikit-learn
pyyaml